+++ PayPalMobile/PayPalMobile.h
@@ -1,7 +1,7 @@
 //
 //  PayPalMobile.h
 //
-//  Version 2.0.5
+//  Version 2.1.0
 //
 //  Copyright (c) 2014, PayPal
 //  All rights reserved.
@@ -63,6 +63,15 @@ extern NSString *const PayPalEnvironmentNoNetwork;
 /// See PayPalEnvironmentProduction, PayPalEnvironmentSandbox, PayPalEnvironmentNoNetwork for more details.
 /// @return applicationCorrelationID Your server will send this to PayPal in a 'Paypal-Application-Correlation-Id' header.
 + (NSString *)applicationCorrelationIDForEnvironment:(NSString *)environment;
+
+/// Delete all previously remembered user data (credit card, email, phone, PayPal account),
+/// for all environments. (See PayPalEnvironmentProduction, PayPalEnvironmentSandbox, PayPalEnvironmentNoNetwork.)
+///
+/// You might choose to call this method, for example, when a user logs out of your app.
+///
+/// @note Do NOT call this method while any PayPal Mobile SDK viewcontroller is active.
+/// @see PayPalConfiguration rememberUser.
++ (void)clearAllUserData;
 
 /// @return The version of the PayPal Mobile SDK in use. Version numbering follows http://semver.org/.
 /// @note Please be sure to include this library version in tech support requests.