+++ SampleApp/PayPal-iOS-SDK-Sample-App/ZZMainViewController.m
@@ -76,10 +76,46 @@
   // Remove our last completed payment, just for demo purposes.
   self.resultText = nil;
   
+  // Note: For purposes of illustration, this example shows a payment that includes
+  //       both payment details (subtotal, shipping, tax) and multiple items.
+  //       You would only specify these if appropriate to your situation.
+  //       Otherwise, you can leave payment.items and/or payment.paymentDetails nil,
+  //       and simply set payment.amount to your total charge.
+  
+  // Optional: include multiple items
+  PayPalItem *item1 = [PayPalItem itemWithName:@"Old jeans with holes"
+                                  withQuantity:2
+                                     withPrice:[NSDecimalNumber decimalNumberWithString:@"84.99"]
+                                  withCurrency:@"USD"
+                                       withSku:@"Hip-00037"];
+  PayPalItem *item2 = [PayPalItem itemWithName:@"Free rainbow patch"
+                                  withQuantity:1
+                                     withPrice:[NSDecimalNumber decimalNumberWithString:@"0.00"]
+                                  withCurrency:@"USD"
+                                       withSku:@"Hip-00066"];
+  PayPalItem *item3 = [PayPalItem itemWithName:@"Long-sleeve plaid shirt (mustache not included)"
+                                  withQuantity:1
+                                     withPrice:[NSDecimalNumber decimalNumberWithString:@"37.99"]
+                                  withCurrency:@"USD"
+                                       withSku:@"Hip-00291"];
+  NSArray *items = @[item1, item2, item3];
+  NSDecimalNumber *subtotal = [PayPalItem totalPriceForItems:items];
+  
+  // Optional: include payment details
+  NSDecimalNumber *shipping = [[NSDecimalNumber alloc] initWithString:@"5.99"];
+  NSDecimalNumber *tax = [[NSDecimalNumber alloc] initWithString:@"2.50"];
+  PayPalPaymentDetails *paymentDetails = [PayPalPaymentDetails paymentDetailsWithSubtotal:subtotal
+                                                                             withShipping:shipping
+                                                                                  withTax:tax];
+
+  NSDecimalNumber *total = [[subtotal decimalNumberByAdding:shipping] decimalNumberByAdding:tax];
+  
   PayPalPayment *payment = [[PayPalPayment alloc] init];
-  payment.amount = [[NSDecimalNumber alloc] initWithString:@"9.95"];
+  payment.amount = total;
   payment.currencyCode = @"USD";
-  payment.shortDescription = @"Hipster t-shirt";
+  payment.shortDescription = @"Hipster clothing";
+  payment.items = items;  // if not including multiple items, then leave payment.items as nil
+  payment.paymentDetails = paymentDetails; // if not including payment details, then leave payment.paymentDetails as nil
 
   if (!payment.processable) {
     // This particular payment will always be processable. If, for